// src/frontend/src/backend.d.ts
export interface NegotiationMessage {
    sender: string;
    message: string;
    timestamp: bigint;
}
export interface QuotationRequest {
    id: string;
    status: QuotationStatus;
    serviceType: ServiceType;
    customer: Principal;
    mobileNumber: string;
    deadline: bigint;
    email: string;
    negotiationHistory: Array<NegotiationMessage>;
    timestamp: bigint;
    projectDetails: string;
}
// ...
export enum QuotationStatus {
    rejected = "rejected",
    accepted = "accepted",
    negotiating = "negotiating",
    pendingCustomerResponse = "pendingCustomerResponse"
}
export enum ServiceType {
    banner = "banner",
    offset = "offset",
    design = "design",
    digital = "digital"
}